Refactor CLI
